Add ADC INPUT and correct bit issue with sending Polybox MCode

+++ src/ArduinoAVR/Repetier/Communication.cpp
@@ -377,7 +377,7 @@ FSTRINGVALUE(Com::tCreationFailed,"Creation failed")
 FSTRINGVALUE(Com::tSDErrorCode,"SD errorCode:")
 #endif // SDSUPPORT
 
-void Com::printPolybox(uint8_t n)
+void Com::printPolybox(int n)
 {
 	print(POLY_SERIAL_SEPARATOR_VALUE);
 	print(n);
